requirements.txt
fastapi
uvicorn
pydantic
transformers
google-generativeai
torch
python-dotenv
torchaudio
scipy
pydub